requirements.txt
requests>=2.28.0
selenium>=4.0.0
beautifulsoup4>=4.11.0
playwright>=1.30.0
aiohttp>=3.8.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
asyncpg>=0.27.0
alembic>=1.9.0
fastapi>=0.95.0
uvicorn>=0.21.0
pydantic>=1.10.0
prometheus-client>=0.16.0
grafana-api>=1.0.0
transformers>=4.27.0
torch>=2.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0
bandit>=1.7.0
safety>=2.3.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
